+++ src/core/parse.js
@@ -1,11 +1,16 @@
 import moduleRequire from './module-require'
 import assign from 'object-assign'
+import { extname } from 'path'
+import debug from 'debug'
 
-export default function (content, context) {
+const log = debug('eslint-plugin-import:parse')
+
+export default function (path, content, context) {
 
   if (context == null) throw new Error('need context to parse properly')
 
-  let { parserOptions, parserPath } = context
+  let { parserOptions } = context
+  const parserPath = getParserPath(path, context)
 
   if (!parserPath) throw new Error('parserPath is required!')
 
@@ -20,4 +25,20 @@ export default function (content, context) {
   const parser = moduleRequire(parserPath)
 
   return parser.parse(content, parserOptions)
+}
+
+function getParserPath(path, context) {
+  const parsers = context.settings['import/parsers']
+  if (parsers != null) {
+    const extension = extname(path)
+    for (let parserPath in parsers) {
+      if (parsers[parserPath].indexOf(extension) > -1) {
+        // use this alternate parser
+        log('using alt parser:', parserPath)
+        return parserPath
+      }
+    }
+  }
+  // default to use ESLint parser
+  return context.parserPath
 }